+++ pkg/model/kubernetes/kubernetes.go
@@ -12,11 +12,11 @@ import (
 	"github.com/Azure/radius/pkg/renderers/containerv1alpha3"
 	"github.com/Azure/radius/pkg/renderers/dapr"
 	"github.com/Azure/radius/pkg/renderers/daprhttproutev1alpha3"
-	"github.com/Azure/radius/pkg/renderers/daprstatestorev1alpha1"
+	"github.com/Azure/radius/pkg/renderers/daprstatestorev1alpha3"
 	"github.com/Azure/radius/pkg/renderers/gateway"
 	"github.com/Azure/radius/pkg/renderers/httproutev1alpha3"
 	"github.com/Azure/radius/pkg/renderers/mongodbv1alpha3"
-	"github.com/Azure/radius/pkg/renderers/rabbitmqv1alpha1"
+	"github.com/Azure/radius/pkg/renderers/rabbitmqv1alpha3"
 	"github.com/Azure/radius/pkg/renderers/redisv1alpha3"
 	"github.com/Azure/radius/pkg/renderers/volumev1alpha3"
 	"github.com/Azure/radius/pkg/resourcekinds"
@@ -27,9 +27,9 @@ func NewKubernetesModel(k8s client.Client) model.ApplicationModel {
 	renderers := map[string]renderers.Renderer{
 		containerv1alpha3.ResourceType:      &dapr.Renderer{Inner: &containerv1alpha3.Renderer{}},
 		daprhttproutev1alpha3.ResourceType:  &daprhttproutev1alpha3.Renderer{},
-		daprstatestorev1alpha1.ResourceType: &daprstatestorev1alpha1.Renderer{StateStores: daprstatestorev1alpha1.SupportedKubernetesStateStoreKindValues},
+		daprstatestorev1alpha3.ResourceType: &daprstatestorev1alpha3.Renderer{StateStores: daprstatestorev1alpha3.SupportedKubernetesStateStoreKindValues},
 		mongodbv1alpha3.ResourceType:        &mongodbv1alpha3.KubernetesRenderer{},
-		rabbitmqv1alpha1.ResourceType:       &rabbitmqv1alpha1.Renderer{},
+		rabbitmqv1alpha3.ResourceType:       &rabbitmqv1alpha3.Renderer{},
 		redisv1alpha3.ResourceType:          &redisv1alpha3.KubernetesRenderer{},
 		httproutev1alpha3.ResourceType:      &httproutev1alpha3.Renderer{},
 		gateway.ResourceType:                &gateway.Renderer{},
