+++ pkg/renderers/rabbitmqv1alpha3/renderer.go
@@ -3,7 +3,7 @@
 // Licensed under the MIT License.
 // ------------------------------------------------------------
 
-package rabbitmqv1alpha1
+package rabbitmqv1alpha3
 
 import (
 	"context"
