+++ test/functional/azure/resources/dapr_pubsub_servicebus_test.go
@@ -13,7 +13,7 @@ import (
 	"github.com/Azure/radius/pkg/radrp/outputresource"
 	"github.com/Azure/radius/pkg/radrp/rest"
 	"github.com/Azure/radius/pkg/renderers/containerv1alpha3"
-	"github.com/Azure/radius/pkg/renderers/daprpubsubv1alpha1"
+	"github.com/Azure/radius/pkg/renderers/daprpubsubv1alpha3"
 	"github.com/Azure/radius/pkg/resourcekinds"
 	"github.com/Azure/radius/test/azuretest"
 	"github.com/Azure/radius/test/validation"
@@ -58,7 +58,7 @@ func Test_DaprPubSubServiceBusManaged(t *testing.T) {
 					{
 						ApplicationName: application,
 						ResourceName:    "pubsub",
-						ResourceType:    daprpubsubv1alpha1.ResourceType,
+						ResourceType:    daprpubsubv1alpha3.ResourceType,
 						OutputResources: map[string]validation.ExpectedOutputResource{
 							outputresource.LocalIDAzureServiceBusTopic: validation.NewOutputResource(outputresource.LocalIDAzureServiceBusTopic, outputresource.TypeARM, resourcekinds.DaprPubSubTopicAzureServiceBus, true, false, rest.OutputResourceStatus{}),
 						},
@@ -116,7 +116,7 @@ func Test_DaprPubSubServiceBusUnmanaged(t *testing.T) {
 					{
 						ApplicationName: application,
 						ResourceName:    "pubsub",
-						ResourceType:    daprpubsubv1alpha1.ResourceType,
+						ResourceType:    daprpubsubv1alpha3.ResourceType,
 						OutputResources: map[string]validation.ExpectedOutputResource{
 							outputresource.LocalIDAzureServiceBusTopic: validation.NewOutputResource(outputresource.LocalIDAzureServiceBusTopic, outputresource.TypeARM, resourcekinds.DaprPubSubTopicAzureServiceBus, false, false, rest.OutputResourceStatus{}),
 						},
